changed data type for pgsql GROUPBY AND DISTINCT removed columns with ::decimal pt2

+++ database/migrations/2020_07_19_155204_create_farmer_totalqty_view.php
@@ -18,7 +18,7 @@ class CreateFarmerTotalqtyView extends Migration
         SELECT 
             posts.user_id AS user_id,
             SUM(posts.crop_quantity::decimal) AS sumCropQty,
-            (SUM(posts.crop_quantity::decimal) * posts.crop_price) AS allCropPrice,
+            (SUM(posts.crop_quantity::decimal) * posts.crop_price::decimal) AS allCropPrice,
             AVG(posts.crop_price::decimal) AS avgCropPrice,
             posts.crop_name AS crop_name
         FROM
